Creating a matrix that displays distances between cities
This program takes number of Indian cities and city names as input from the user. It then scrapes data from the website www.distancefromto.net to find out the distances between all the possible combinations of the cities. At the end it displays the distances between all possible combinations in the form of a dataframe.